Fix Windows build: u32 slice index and platform-specific anyhow import
- Fix windows_tap.rs: cast data_len to usize before saturating_sub for slice indexing
- Fix tap.rs: split anyhow import to be platform-specific to avoid unused import warning

+++ src/windows_tap.rs
@@ -63,7 +63,7 @@ fn find_tap_guid(device_name: &str) -> Result<String> {
 
                 if RegQueryValueExA(subkey, name_key.as_ptr(), std::ptr::null_mut(), std::ptr::null_mut(), data.as_mut_ptr(), &mut data_len) == 0 {
                     let name = if data_len > 0 {
-                        std::str::from_utf8(&data[..data_len.saturating_sub(1)]).unwrap_or("").trim_matches('\0')
+                        std::str::from_utf8(&data[..(data_len as usize).saturating_sub(1)]).unwrap_or("").trim_matches('\0')
                     } else {
                         ""
                     };
